requirements.txt
slack_sdk>=3.0
slackeventsapi>=2.1.0
Flask>=1.1.1
certifi
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
beautifulsoup4
gunicorn
oauth2client
gspread